fix: 로그인 기능을 signin.js에 병합
